move problematic hash functions to separate file
